refactored page sections and moved styles below components
