+++ components/MessageButton.js
@@ -3,12 +3,12 @@ import styled from 'styled-components';
 
 import { orangeButtonStyles } from './styled/global';
 
-const StyledButton = styled.a`
-  ${orangeButtonStyles}
-`;
-
 const MessageButton = () => {
   return <StyledButton href="/#contact">Send Me a Message</StyledButton>;
 };
+
+const StyledButton = styled.a`
+  ${orangeButtonStyles}
+`;
 
 export default MessageButton;
